api/requirements.txt
# Make sure that all your apps (folders) will have own requirements

# RabbitMQ
# Async driver https://aio-pika.readthedocs.io/en/latest/
aio-pika==9.4.1
# Sync driver https://www.rabbitmq.com/tutorials/tutorial-one-python
pika


fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic-settings==2.2.1

# OpenTelemetry instrumentation
opentelemetry-api==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-instrumentation-requests==0.45b0
requests==2.31.0